refactor: TestDiscordAgent を discord-agent-test-helpers に集約
spec と unit test に重複していた TestDiscordAgent・createSessionPort・
createAgent を spec/agent/discord/discord-agent-test-helpers.ts に切り出し、
両ファイルから共有する。

Closes #791
