+++ packages/agent/src/discord/discord-agent.test.ts
@@ -1,109 +1,8 @@
 /* oxlint-disable max-lines, max-lines-per-function -- テストファイルはケース数に応じて長くなるため許容 */
 import { afterEach, describe, expect, mock, test } from "bun:test";
 
-import type { OpencodeSessionPort } from "@vicissitude/shared/types";
-
-import {
-	createContextBuilder,
-	createProfile,
-	createSessionStore,
-} from "../../../../spec/agent/runner-test-helpers.ts";
-import { createMockLogger } from "../../../../spec/test-helpers.ts";
+import { createAgent } from "../../../../spec/agent/discord/discord-agent-test-helpers.ts";
 import type { AgentRunner } from "../runner.ts";
-import { DiscordAgent, type DiscordAgentDeps } from "./discord-agent.ts";
-
-// ─── テスト用 DiscordAgent サブクラス ─────────────────────────────
-
-class TestDiscordAgent extends DiscordAgent {
-	sleepSpy: ((ms: number) => Promise<void>) | null = null;
-
-	/** pendingCompaction フラグへの公開アクセス */
-	get isPendingCompaction(): boolean {
-		return (this as unknown as { pendingCompaction: boolean }).pendingCompaction;
-	}
-
-	/** lastActivityAt への公開アクセス */
-	get currentLastActivityAt(): number | null {
-		return (this as unknown as { lastActivityAt: number | null }).lastActivityAt;
-	}
-
-	/** lastChannelId への公開アクセス */
-	get currentLastChannelId(): string | null {
-		return (this as unknown as { lastChannelId: string | null }).lastChannelId;
-	}
-
-	/** compactionGapMs への公開アクセス */
-	get currentCompactionGapMs(): number {
-		return (this as unknown as { compactionGapMs: number }).compactionGapMs;
-	}
-
-	/** rotationGapMs への公開アクセス */
-	get currentRotationGapMs(): number {
-		return (this as unknown as { rotationGapMs: number }).rotationGapMs;
-	}
-
-	/** nowProvider への公開アクセス */
-	get currentNowProvider(): () => number {
-		return this.nowProvider;
-	}
-
-	readonly requestSessionRotationMock = mock((): Promise<void> => Promise.resolve());
-
-	// oxlint-disable-next-line no-useless-constructor -- DiscordAgentDeps 型を受け取るために必要
-	constructor(deps: DiscordAgentDeps) {
-		super(deps);
-	}
-
-	override requestSessionRotation(): Promise<void> {
-		const result: Promise<void> = this.requestSessionRotationMock();
-		return result;
-	}
-
-	/** テスト用: polling loop を起動しない（pendingCompaction フラグが消費されるのを防ぐ） */
-	override ensurePolling(): void {}
-
-	protected override sleep(ms: number): Promise<void> {
-		if (this.sleepSpy) return this.sleepSpy(ms);
-		return super.sleep(ms);
-	}
-}
-
-// ─── ヘルパー ─────────────────────────────────────────────────────
-
-function createSessionPort(): OpencodeSessionPort {
-	return {
-		createSession: mock(() => Promise.resolve("session-1")),
-		sessionExists: mock(() => Promise.resolve(false)),
-		prompt: mock(() => Promise.resolve({ text: "", tokens: undefined })),
-		promptAsync: mock(() => Promise.resolve()),
-		promptAsyncAndWatchSession: mock(() => Promise.resolve({ type: "idle" as const })),
-		waitForSessionIdle: mock(() => Promise.resolve({ type: "idle" as const })),
-		deleteSession: mock(() => Promise.resolve()),
-		summarizeSession: mock(() => Promise.resolve()),
-		close: mock(() => {}),
-	} as unknown as OpencodeSessionPort;
-}
-
-function createAgent(
-	opts: {
-		nowProvider?: () => number;
-		conversationBreak?: { compactionGapMs?: number; rotationGapMs?: number };
-	} = {},
-): TestDiscordAgent {
-	const agent = new TestDiscordAgent({
-		guildId: "guild-1",
-		profile: createProfile(),
-		sessionStore: createSessionStore() as never,
-		contextBuilder: createContextBuilder(),
-		logger: createMockLogger(),
-		sessionPort: createSessionPort(),
-		sessionMaxAgeMs: 86_400_000,
-		nowProvider: opts.nowProvider,
-		conversationBreak: opts.conversationBreak,
-	});
-	agent.sleepSpy = () => Promise.resolve();
-	return agent;
-}
 
 const activeRunners = new Set<AgentRunner>();
